cloud_functions/requirements.txt
Flask
pandas
numpy
google-cloud-storage
google-cloud-bigquery
lxml